backend/requirements.txt
asgiref==3.3.1
cffi==1.14.4
cryptography==3.2.1
Django==3.1.4
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
ecdsa==0.14.1
pyasn1==0.4.8
pycparser==2.20
python-jose==3.2.0
pytz==2020.4
rsa==4.6
six==1.15.0
sqlparse==0.4.1
